+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [0.0.2]-2018-12-21
+
+### Added
+ - TypeScript 支持
+ - 日志配置添加过滤headr字段配置`filterHeaders`
+
+```
+logCofig: {
+	logFilePath: 'string',
+	pm2: true,
+	filterHeaders: []
+	...// log4js config [https://log4js-node.github.io/log4js-node/](https://log4js-node.github.io/log4js-node/)
+}
+```
+
+
 ## [0.0.1]-2018-12-18
 
 ### Added